DSView.desktop: remove known version in desktop file
Fix error: value "0.96" for key "Version" in group "Desktop Entry"
is not a known version.

Version of the Desktop Entry Specification that the desktop entry
conforms with. Entries that confirm with this version of the
specification should use 1.1. Note that the version field is not
required to be present, so I removed it.
See also: https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html

Signed-off-by: Huang Rui <vowstar@gmail.com>