Decoder sdcard: look at the transmission bit for command or response decoding
Description of the bug fix:
When a command is not supported by the sdcard, the command is silently
dropped with no response. The host will then either reemit the same command
or give up. This causes the previous decoder implementation to go out of
sync and wrongly decode the next message as a response even though it's not.

This patch looks at the transmission bit instead to make sure the
decoding is correct.