common/minizip/*: fix OF prototypes
Because the sys-libs/zlib header pollutes the macro namespace,
leading to build failures. Under the gentoo operating system,
the macro definition of OF becomes _Z_OF in zconf.h, but the
recently introduced minizip has introduced a large number of OF,
leading to build fail. Therefore, the prototypes of OF are added
so that they can be successfully built under various operating
systems.

Bug: https://bugs.gentoo.org/383179
Signed-off-by: Huang Rui <vowstar@gmail.com>